BZ-1166011 - SessionInfo primary key is Integer

+++ kie-infinispan/drools-infinispan-persistence/src/main/java/org/drools/persistence/info/EntityHolder.java
@@ -20,7 +20,7 @@ public class EntityHolder {
 	@Field
 	private String type;
 	@Field
-	private Integer sessionInfoId;
+	private Long sessionInfoId;
 	@Field
 	private Integer sessionInfoVersion;
 	@Field
